moved some methods to Ui class
